+++ galasa-ui/src/components/test-runs/test-run-details/TestRunDetails.tsx
@@ -43,9 +43,11 @@ import {
   TEST_RUN_TAB_NAMES,
 } from '@/utils/constants/common';
 import { NotificationType } from '@/utils/types/common';
-import { TreeNodeData } from '@/utils/functions/artifacts';
+import { TreeNodeData, FolderNode } from '@/utils/functions/artifacts';
 import TestRunsSearch from '../TestRunsSearch';
 import { getExistingTagObjects } from '@/actions/runsAction';
+import { checkForZosTerminalFolderStructure } from '@/utils/3270/checkFor3270FolderStructure';
+import { cleanArtifactPath } from '@/utils/artifacts';
 
 interface TestRunDetailsProps {
   runId: string;
@@ -62,6 +64,11 @@ const TestRunDetails = ({ runId, runDetailsPromise }: TestRunDetailsProps) => {
   const [run, setRun] = useState<RunMetadata>();
   const [methods, setMethods] = useState<TestMethod[]>([]);
   const [artifacts, setArtifacts] = useState<ArtifactIndexEntry[]>([]);
+  const [artifactsTreeData, setArtifactsTreeData] = useState<FolderNode>({
+    name: '',
+    isFile: false,
+    children: {},
+  });
   const [artifactsLoaded, setArtifactsLoaded] = useState(false);
   const [artifactsLoading, setArtifactsLoading] = useState(false);
   const [artifactsError, setArtifactsError] = useState<string | null>(null);
@@ -150,9 +157,9 @@ const TestRunDetails = ({ runId, runDetailsPromise }: TestRunDetailsProps) => {
         duration:
           runDetails.testStructure?.startTime && runDetails.testStructure?.endTime
             ? getIsoTimeDifference(
-                runDetails.testStructure?.startTime,
-                runDetails.testStructure?.endTime
-              )
+              runDetails.testStructure?.startTime,
+              runDetails.testStructure?.endTime
+            )
             : '-',
         tags: runDetails.testStructure?.tags || [],
       };
@@ -176,6 +183,69 @@ const TestRunDetails = ({ runId, runDetailsPromise }: TestRunDetailsProps) => {
       const runArtifacts = await response.json();
       setArtifacts(runArtifacts);
       setArtifactsLoaded(true);
+
+      // Build tree structure to check for 3270 terminal folder
+      if (runArtifacts.length > 0) {
+        const root: FolderNode = { name: '', isFile: false, children: {} };
+
+        runArtifacts.forEach((artifact: ArtifactIndexEntry) => {
+          const rawPath = artifact.path ?? '';
+          const cleanedPath = cleanArtifactPath(rawPath);
+          let segments = cleanedPath.split('/').filter((seg) => seg !== '');
+
+          const segmentValue = segments[0]?.toLocaleLowerCase();
+          if (segmentValue === 'artifact' || segmentValue === 'artifacts') {
+            segments = segments.slice(1);
+          }
+
+          if (segments.length > 0) {
+            let currentNode: FolderNode = root;
+            segments.forEach((segment, idx) => {
+              const isLast = idx === segments.length - 1;
+
+              if (isLast) {
+                currentNode.children[segment] = {
+                  name: segment,
+                  runId: artifact.runId ?? '',
+                  url: artifact.path ?? '',
+                  isFile: true,
+                  children: {},
+                };
+              } else {
+                const existing = currentNode.children[segment];
+
+                if (!existing) {
+                  currentNode.children[segment] = {
+                    name: segment,
+                    isFile: false,
+                    children: {},
+                  };
+                  currentNode = currentNode.children[segment] as FolderNode;
+                } else if (existing.isFile) {
+                  currentNode.children[segment] = {
+                    name: segment,
+                    isFile: false,
+                    children: {},
+                  };
+                  currentNode = currentNode.children[segment] as FolderNode;
+                } else {
+                  currentNode = existing as FolderNode;
+                }
+              }
+            });
+          }
+        });
+
+        // Store the built tree for use in ArtifactsTab
+        setArtifactsTreeData(root);
+
+        // Check for 3270 terminal folder structure
+        checkForZosTerminalFolderStructure(
+          root,
+          setZos3270TerminalFolderExists,
+          setZos3270TerminalData
+        );
+      }
     } catch (err: unknown) {
       console.error('Error loading artifacts:', err);
       setArtifactsError(err instanceof Error ? err.message : 'Failed to load artifacts');
@@ -229,13 +299,12 @@ const TestRunDetails = ({ runId, runDetailsPromise }: TestRunDetailsProps) => {
     loadRunDetails();
   }, [run, runDetailsPromise, extractRunDetails]);
 
-  // If artifacts tab is selected in URL on initial load, fetch artifacts
+  // Load artifacts immediately when run details are loaded to check for 3270 terminal folder
   useEffect(() => {
-    const artifactsTabIndex = TEST_RUN_PAGE_TABS.indexOf(TEST_RUN_TAB_NAMES.ARTIFACTS);
-    if (selectedTabIndex === artifactsTabIndex && !artifactsLoaded && !artifactsLoading && run) {
+    if (run && !artifactsLoaded && !artifactsLoading && !artifactsError) {
       loadArtifacts();
     }
-  }, [selectedTabIndex, artifactsLoaded, artifactsLoading, run, loadArtifacts]);
+  }, [run, artifactsLoaded, artifactsLoading, artifactsError, loadArtifacts]);
 
   // If log tab is selected in URL on initial load, fetch logs
   useEffect(() => {
@@ -492,6 +561,7 @@ const TestRunDetails = ({ runId, runDetailsPromise }: TestRunDetailsProps) => {
               <TabPanel>
                 <ArtifactsTab
                   artifacts={artifacts}
+                  artifactsTreeData={artifactsTreeData}
                   runId={runId}
                   runName={run?.runName || ''}
                   isLoadingArtifacts={artifactsLoading}
